first pass at multiprotocol support
